fix: cast Tax value to long in PayLicenseCommandHandler and clean up Price value object calculations

+++ src/domain/CodeDesignPlus.Net.Microservice.Licenses.Application/License/Commands/PayLicense/PayLicenseCommandHandler.cs
@@ -51,7 +51,7 @@ public class PayLicenseCommandHandler(ILicenseRepository repository, IUserContex
         payRequest.Transaction.Order.Tax = new Amount
         {
             Currency = price.Currency.Code,
-            Value = price.Tax
+            Value = (long)price.Tax
         };
 
         payRequest.Transaction.Order.Amount = new Amount
